+++ components/accounts/accounts_context_menu.h
@@ -3,6 +3,6 @@
 #include <string>
 #include "../data.h"
 
-void RenderAccountContextMenu(AccountData& account, const std::string& unique_context_menu_id);
+void RenderAccountContextMenu(AccountData &account, const std::string &unique_context_menu_id);
 
 void LaunchBrowserWithCookie(const AccountData &account);
