+++ components/accounts/accounts_join_ui.cpp
@@ -8,7 +8,7 @@
 #include <stdexcept>
 #include <utility>
 
-#include "roblox_api.h"
+#include "roblox.h"
 #include "threading.h"
 #include "../data.h"
 #include "../../ui.h"
