+++ components/accounts/accounts_tab.cpp
@@ -16,7 +16,7 @@
 #include "webview.hpp"
 
 #include "system/threading.h"
-#include "network/roblox_api.h"
+#include "network/roblox.h"
 #include "core/time_utils.h"
 #include "core/logging.hpp"
 #include "core/status.h"
