+++ components/components.h
@@ -9,20 +9,20 @@ void RenderAccountsTable(std::vector<AccountData> &, const char *, float);
 bool RenderMainMenu();
 
 struct PublicServerInfo {
-    std::string jobId;
-    int currentPlayers = 0;
-    int maximumPlayers = 0;
-    double averagePing = 0.0;
-    double averageFps = 0.0;
+	std::string jobId;
+	int currentPlayers = 0;
+	int maximumPlayers = 0;
+	double averagePing = 0.0;
+	double averageFps = 0.0;
 };
 
 struct GameInfo {
-    std::string name;
-    uint64_t universeId = 0;
-    uint64_t placeId = 0;
-    int playerCount = 0;
-    int upVotes = 0;
-    int downVotes = 0;
-    std::string creatorName;
-    bool creatorVerified = false;
+	std::string name;
+	uint64_t universeId = 0;
+	uint64_t placeId = 0;
+	int playerCount = 0;
+	int upVotes = 0;
+	int downVotes = 0;
+	std::string creatorName;
+	bool creatorVerified = false;
 };
