+++ components/friends/friends_actions.cpp
@@ -1,5 +1,5 @@
 #include "friends_actions.h"
-#include "network/roblox_api.h"
+#include "network/roblox.h"
 #include "core/status.h"
 #include <algorithm>
 #include <vector>
