+++ components/friends/friends_actions.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <atomic>
 
-#include "network/roblox_api.h"
+#include "network/roblox.h"
 #include "../data.h"
 
 namespace FriendsActions {
