+++ components/friends/friends_tab.cpp
@@ -9,7 +9,7 @@
 
 
 #include "../data.h"
-#include "network/roblox_api.h"
+#include "network/roblox.h"
 #include "system/launcher.hpp"
 #include "system/threading.h"
 #include "./friends_actions.h"
