+++ components/games/games_tab.cpp
@@ -12,7 +12,7 @@
 
 #include "../components.h"
 #include "system/launcher.hpp"
-#include "network/roblox_api.h"
+#include "network/roblox.h"
 #include "core/status.h"
 #include "ui/modal_popup.h"
 #include "../../ui.h"
