+++ components/history/log_parser.h
@@ -3,8 +3,6 @@
 #include "log_types.h"
 #include <string>
 
-// Parses a single log file and populates the LogInfo struct.
-void parseLogFile(LogInfo& logInfo);
+void parseLogFile(LogInfo &logInfo);
 
-// Returns the path to the Roblox logs folder.
 std::string logsFolder();
