+++ components/servers/servers_tab.cpp
@@ -16,7 +16,7 @@
 #include <cstdio>
 
 #include "../components.h"
-#include "network/roblox_api.h"
+#include "network/roblox.h"
 #include "core/status.h"
 #include "system/launcher.hpp"
 #include "ui/modal_popup.h"
