+++ utils/system/main_thread.h
@@ -4,23 +4,22 @@
 #include <mutex>
 
 namespace MainThread {
-    using Task = std::function<void()>;
-    inline std::deque<Task> tasks;
-    inline std::mutex mtx;
+	using Task = std::function<void()>;
+	inline std::deque<Task> tasks;
+	inline std::mutex mtx;
 
-    inline void Post(Task t) {
-        std::lock_guard<std::mutex> lock(mtx);
-        tasks.push_back(std::move(t));
-    }
+	inline void Post(Task t) {
+		std::lock_guard<std::mutex> lock(mtx);
+		tasks.push_back(std::move(t));
+	}
 
-    inline void Process() {
-        std::deque<Task> toRun;
-        {
-            std::lock_guard<std::mutex> lock(mtx);
-            toRun.swap(tasks);
-        }
-        for (auto &t : toRun) {
-            t();
-        }
-    }
+	inline void Process() {
+		std::deque<Task> toRun; {
+			std::lock_guard<std::mutex> lock(mtx);
+			toRun.swap(tasks);
+		}
+		for (auto &t: toRun) {
+			t();
+		}
+	}
 }
